@@ -8,21 +8,20 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2",
-        "codersstudio/crud": "dev-master",
-        "fideloper/proxy": "^4.0",
-        "laravel/framework": "^6.2",
-        "laravel/tinker": "^2.0",
-        "laravel/ui": "^1.1",
-        "spatie/laravel-medialibrary": "^7.0.0",
-        "spatie/laravel-permission": "^3.6"
+        "php": "^8.1",
+        "imjonos/laravel-crud-generator": "*",
+        "fideloper/proxy": "^4.4",
+        "laravel/framework": "^8.12",
+        "laravel/tinker": "^2.5",
+        "laravel/ui": "^3.2",
+        "spatie/laravel-permission": "^4.0"
     },
     "require-dev": {
-        "facade/ignition": "^1.4",
-        "fzaninotto/faker": "^1.4",
-        "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^3.0",
-        "phpunit/phpunit": "^8.0"
+        "facade/ignition": "^2.5",
+        "fakerphp/faker": "*",
+        "mockery/mockery": "^1.4.2",
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.3.3"
     },
     "config": {
         "optimize-autoloader": true,
@@ -39,7 +38,7 @@
             "App\\": "app/"
         },
         "classmap": [
-            "database/seeds",
+            "database/seeders",
             "database/factories"
         ]
     },
