@@ -1,8 +1,8 @@
 <?php
 /**
- * CodersStudio 2019
- * https://coders.studio
- * info@coders.studio
+ * Eugeny Nosenko 2021
+ * https://toprogram.ru
+ * info@toprogram.ru
  */
 
 namespace App\Http\Requests\Admin\Comment;
@@ -11,7 +11,7 @@ use Illuminate\Foundation\Http\FormRequest;
 
 /**
  * Class EditRequest
- * @package CodersStudio\CRUD
+ * @package Nos\CRUD
  */
 class EditRequest extends FormRequest
 {
