@@ -1,21 +1,38 @@
 <?php
 /**
- * CodersStudio 2019
- * https://coders.studio
- * info@coders.studio
+ * Eugeny Nosenko 2021
+ * https://toprogram.ru
+ * info@toprogram.ru
  */
+ 
+namespace Database\Factories;
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
-use App\Comment;
-use Illuminate\Support\Str;
-use Faker\Generator as Faker;
+use App\Models\Comment;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Comment::class, function (Faker $faker) {
-    return [
-        'user_name' => $faker->word,
-        'publish' => $faker->boolean(),
-        'comment' => $faker->word,
-        'post_id' => App\Post::first()->id,
+class CommentFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Comment::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'user_name' => $this->faker->word,
+        'publish' => $this->faker->boolean(),
+        'comment' => $this->faker->word,
+        'post_id' => \App\Models\Post::first()->id,
         
-    ];
-});
+        ];
+    }
+}
+
