@@ -7,9 +7,7 @@
 
 require('./bootstrap');
 
-window.Vue = require('vue');
-
-require('./vendor/codersstudio/crud/index');
+window.Vue = require('vue').default;
 /**
  * The following block of code may be used to automatically register your
  * Vue components. It will recursively scan this directory for the Vue
@@ -21,13 +19,15 @@ require('./vendor/codersstudio/crud/index');
 // const files = require.context('./', true, /\.vue$/i)
 // files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
 
-Vue.component('example-component', require('./components/ExampleComponent.vue').default);
+//Vue.component('example-component', require('./components/ExampleComponent.vue').default);
 
 /**
  * Next, we will create a fresh Vue application instance and attach it to
  * the page. Then, you may begin adding components to this application
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
+
+require('./vendor/nos/crud/index');
 
 const app = new Vue({
     el: '#app',
