@@ -1,7 +1,6 @@
-@extends('codersstudio.crud::layouts.app')
+@extends('nos.crud::layouts.app')
 
 @section('title', trans('crud.post.title'))
-
 
 @section('content')
     <p>This is my body content.</p>
